+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/stub/RobotScalarVariableStubElement.java
@@ -13,6 +13,7 @@ import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotScalarVariable;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotVariableBodyId;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.impl.RobotScalarVariableImpl;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.stub.index.VariableNameIndex;
+import dev.xeonkryptos.xeonrobotframeworkplugin.util.VariableNameUtil;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
 
@@ -68,8 +69,7 @@ public class RobotScalarVariableStubElement extends IStubElementType<RobotScalar
     public void indexStub(@NotNull RobotScalarVariableStub stub, @NotNull IndexSink sink) {
         String variableName = stub.getVariableName();
         if (variableName != null) {
-            String variableNameInLowerCase = variableName.toLowerCase();
-            sink.occurrence(VariableNameIndex.KEY, variableNameInLowerCase);
+            VariableNameUtil.INSTANCE.computeVariableNameVariants(variableName).forEach(variant -> sink.occurrence(VariableNameIndex.KEY, variant));
         }
     }
 }