+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/stub/index/VariableNameIndex.java
@@ -6,6 +6,7 @@ import com.intellij.psi.stubs.StringStubIndexExtension;
 import com.intellij.psi.stubs.StubIndex;
 import com.intellij.psi.stubs.StubIndexKey;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotVariable;
+import dev.xeonkryptos.xeonrobotframeworkplugin.util.VariableNameUtil;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -29,7 +30,10 @@ public class VariableNameIndex extends StringStubIndexExtension<RobotVariable> {
 
     public Collection<RobotVariable> getVariables(@NotNull String unwrappedVariableName, @NotNull Project project, @Nullable GlobalSearchScope scope) {
         StubIndexKey<String, RobotVariable> stubIndexKey = getKey();
-        String variableNameInLowerCase = unwrappedVariableName.toLowerCase();
-        return StubIndex.getElements(stubIndexKey, variableNameInLowerCase, project, scope, RobotVariable.class);
+        return VariableNameUtil.INSTANCE.computeVariableNameVariants(unwrappedVariableName)
+                                        .stream()
+                                        .flatMap(variant -> StubIndex.getElements(stubIndexKey, variant, project, scope, RobotVariable.class).stream())
+                                        .distinct()
+                                        .toList();
     }
 }